Conflicts:
	.gitignore
	app/mailers/user_mailer.rb
	app/models/contact.rb
	app/views/contacts/_form.html.erb
	app/views/contacts/edit.html.erb
	app/views/contacts/index.html.erb
	app/views/contacts/show.html.erb
	app/views/feedbacks/_form.html.erb
	app/views/search_queries/show.html.erb
	app/views/search_records/show.html.erb
	app/views/user_mailer/send_change_of_syndicate_notification_to_sc.html.erb
	app/views/user_mailer/volunteer.html.erb
	app/views/user_mailer/website.html.erb
	app/views/userid_details/index.html.erb
	lib/tasks/mortality_csvs.rake

+++ app/controllers/csvfiles_controller.rb
@@ -26,28 +26,28 @@ class CsvfilesController < ApplicationController
     @csvfile.userid = session[:userid]   if params[:csvfile][:userid].nil?
     @csvfile.file_name = @csvfile.csvfile.identifier
     if params[:commit] == "Replace"
-      #on a replace make sure its the same file_name
-      if session[:file_name] == @csvfile.file_name
-        #set up to allow the file save to occur in check_for_existing_place
-        batch = PhysicalFile.where(userid: @csvfile.userid, file_name: @csvfile.file_name).first
-        unless batch.nil?
-          batch.update_attributes(:base => true,:file_processed => false)
-        else
-          batch = PhysicalFile.new(:base => true,:file_processed => false, :userid =>@csvfile.userid , :file_name => @csvfile.file_name)
-          batch.save
-        end
-      else
+      name_ok = @csvfile.check_name(session[:file_name])
+      if !name_ok
         flash[:notice] = 'The file you are replacing must have the same name'
         session.delete(:file_name)
         redirect_to :back
         return
+      else
+        setup = @csvfile.setup_batch
+        if !setup[0]
+          flash[:notice] = setup[1]
+          session.delete(:file_name)
+          redirect_to :back
+          return
+        else
+          batch = setup[1]   
+        end
       end
-      session.delete(:file_name)
     end
     #lets check for existing file, save if required
-    proceed = @csvfile.check_for_existing_unprocessed_file
+    proceed = @csvfile.check_for_existing_file
     @csvfile.save if proceed
-    if @csvfile.errors.any? || !proceed
+    if @csvfile.errors.any? 
       flash[:notice] = "The upload with file name #{@csvfile.file_name} was unsuccessful because #{@csvfile.errors.messages}"
       get_userids_and_transcribers
       redirect_to :back
@@ -88,7 +88,7 @@ class CsvfilesController < ApplicationController
       if params[:commit] == 'Process'
         @csvfile = Csvfile.where(_id: params[:id]).first
         if @csvfile.nil?
-          flash[:notice] = "There was no file to process; did you perhaps double click or reload the process page?"
+          flash[:notice] = "There was no file to process; did you perhaps double click or reload the process page? Talk to your coordinator if this continues"
           redirect_to action: :new
           return
         end
@@ -99,9 +99,24 @@ class CsvfilesController < ApplicationController
           flash[:notice] =  "The csv file #{ @csvfile.file_name} is being checked. You will receive an email when it has been completed."
         when params[:csvfile][:process]  == "Process tonight"
           batch = PhysicalFile.where(:userid => @csvfile.userid, :file_name => @csvfile.file_name).first
+          if batch.nil?
+            flash[:notice] = "There was no file to put into the queue; did you perhaps double click or reload the process page? Talk to your coordinator if this continues"
+            logger.warn("CSV_FAILURE: No file for #{session[:userid]}")
+            @csvfile.delete
+            redirect_to action: :new
+            return
+          end
           batch.add_file("base")
           flash[:notice] =  "The file has been placed in the queue for overnight processing"
         when params[:csvfile][:process]  == "As soon as you can"
+          batch = PhysicalFile.where(:userid => @csvfile.userid, :file_name => @csvfile.file_name,:waiting_to_be_processed => true).first
+          if batch.present?
+            flash[:notice] = "Your file is currently waiting to be processed. It cannot be processed this way now"
+            logger.warn("CSV_FAILURE: Attempt to double process #{@csvfile.userid} #{@csvfile.file_name}")
+            @csvfile.delete
+            redirect_to action: :new
+            return
+          end
           pid1 = Kernel.spawn("rake build:freereg_update[#{range},\"search_records\",\"change\"]")
           flash[:notice] =  "The csv file #{ @csvfile.file_name} is being processed. You will receive an email when it has been completed."
         else
@@ -153,6 +168,10 @@ class CsvfilesController < ApplicationController
   def download
     @role = session[:role]
     @freereg1_csv_file = Freereg1CsvFile.id(params[:id]).first
+    if  @freereg1_csv_file.nil?
+      flash[:notice] =  "There is no batch to download."
+      redirect_to :back and return
+    end
     errors =  @freereg1_csv_file.check_file
     if errors[0]
       log_messenger("BATCH_ERRORS #{errors[1]}")
